[p9: no sense] (.spec) C++ standard increased for std::optional from the new APT
In p9, however, it makes no sense to cherry-pick some changes in apt
and packagekit and increase the C++ standard, because std::optional
appears only in the big change in ALT's apt API (alt72).
(Unlike std::experimental::optional, understood with c++14,
which appears earlier, in alt70, and so is already present in p9.)

And it'd be simpler just to copy the releases of apt and packagekit
from Sisyphus or p10.