aptcc: mark methods as override to avoid coding errors when API changes
This helps to avoid errors when the library API (and/or some base
classes) change.

The earlier we mark methods override, the easier it will be to catch
dependent changes required after API changes in future.

The "override" mark (keyword) is available since C++11.

(initially just 1 mark was done in a2386d38298a30039df2e2fcf9f7fd9595ae5b2a)
(= 1.2.3-alt2~5)

I don't quite understand why only one place to put the mark was found
initially. -Werror=suggest-override should have done the complete
work (for the use of this flag, see the next changes in the series),
but obviously it didn't. So, it may not work as expected (maybe for
methods from external system headers). If so, I don't quite understand
how then all the numerous methods to mark were actually found when
this patch was enhanced. Maybe this was possible when it was in done
in p9 due to an older gcc behaving differently or the old build
system (Makefile.am) passes the includedirs differently from meson,
which is used currently.