(.spec) move to the beginning: %%set_verify_elf_method strict
This is to avoid it being not seen by a maintainer.

Leave the statement that has this effect only in one place in the spec,
namely, at the beginning. This way it won't be confusing for the maintainer
where to control it.

(cherry picked from commit 36f44649e4fbb0fa5275a9d29554837b1485a824)
(= 1.2.5-alt7~3)