apt: access lists of owned files via rpm
The new (restored) code is the one reverted in:

6db26c8e5 Cut functions depending on access to rpm via apt

In this commit, it is applied on top of the upstream code
in order to see which changes were needed to switch the implementation
from deb to directly accessing rpm. (emitPackageFilesLocal()
could also be implemented via rpm or apt (TODO), but it had been dropped
in ALT before these changes, so this commit doesn't demonstrate
a possible implementation; it remains via deb.)

Probably, it doesn't make sense to compile this mix.