APT backend: drop an argument from MarkInstall() that is not used inside it
Not sure when the last arg appeared here, but it made no sense,
because it is "depth", a left-over from the internals of
the recursive implementation of the function. It has a default value,
and is not used in any way inside this function now. We are going to
drop it from APT's API and ABI completely.

As can be seen from the patch of

8a3b32ef9 Fix building with libapt from ALT

this argument was not the last one in the upstream API, but since
PackageKit wanted to pass the last one (related to auto-marks),
this one couldn't be dropped in the upstream code, although
it made no big sense:

@@ -483,17 +442,17 @@ bool AptCacheFile::tryToInstall(pkgProblemResolver &Fix,
     // non-indicative of whether or not the user explicitly wants this package to be
     // installed or simply wants it to be updated.
     const bool fromUser = preserveAuto ? !(State.Flags & pkgCache::Flag::Auto) : true;
     // FIXME: this is ignoring the return value. OTOH the return value means little to us
     //   since we run markinstall twice, once without autoinst and once with.
     //   We probably should change the return value behavior and have the callee decide whether to
     //   error out or call us again with autoinst. This however is further complicated by us
     //   having protected, so we'd have to lift protection before this?
-    GetDepCache()->MarkInstall(Pkg, autoInst, 0, fromUser);
+    GetDepCache()->MarkInstall(Pkg, (fromUser ? pkgDepCache::AutoMarkFlag::Manual : pkgDepCache::AutoMarkFlag::Auto), autoInst, 0);