also sanitize Suite and Component similarly to Origin
Perhaps, in Debian, these strings don't have bad chars, but other repo
publishers (or apt-rpm repos) may not follow this rule strictly.
So, it's better to be on the safer side.

This change was a part of the commit:

Change processing of origin for ALT repositores

(cherry picked from commit 2cfd5277a4763439c93a91ad5288ec184ec665b0)