unify the processing of the shortened ALT Origin-Suite with the general scheme
The resulting string for ALT must remain essentially the same as it was,
but with the usual PK's delimiters at new places instead of the parentheses.
(Origin and Suite are shortened for ALT by avoiding repeating words, and
the resulting string must be approx. as short as it was, namely 1 char shorter.)

Enhances: 2cfd5277a Change processing of origin for ALT repositores