updateInterface: don't track whether already errorEmitted (non-significant in ALT)
The whole updateInterface func is not going to be used in ALT at all.
(This brings the code closer to the previous version so that further
patches can be applied easier.)

The errorEmitted var was used to pass information from
updateInterface() to the caller about whether an error was already
signalled to the UI or the caller has to do it itself. (AFAIU)

This reverts the essence of PACKAGEKIT_1_2_5~3:
15fe84f04 (aptcc: Emit only the first, likely highest-quality error when dpkg fails, 2022-02-17)

The whole updateInterface func will be dropped altogether in 1.2.5.0.0.30-alt1~4:
facf631cc (drop unused AptIntf::updateInterface() (employed for parsing dpkg output), 2023-07-05)

Follows 1.1.12-alt1~3 (i.e., makes the code similar to what was
after that patch in ALT):
4f162d329 (Remove debconf support, 2018-06-15)