AptCacheFile: always use GetPkgRecords() for access and check it right there
(This is an equiv.rewrite.)

Firstly, this implements the idea that m_packageRecords should be
hidden and built as needed when it is used. (Similarly to how it was
done in ALT's apt with Cache, DCache, etc.; and it could be made
private here, too, to enforce this idea.)

Secondly, this makes it clear in the code that the thing that is being
accessed has been checked for validity right before. (Easier to read
and understand.)

(This change is to be included in release 1.2.5.0.0.30-alt4)