(.spec) don't set a C++ std in optflags; get rid of a lot of useless C warnings
The warnings were from the C compiler, which got a C++ option.

C++ std was needed because APT headers used new C++ features. (Setting
it here is a workaround instead of having it in APT pkg-config.)

Now, it's set in meson.build to a value which is large enough for APT headers.

Since PACKAGEKIT_1_2_5~4
25de1eba8 (trivial: aptcc: Build with older Meson versions, 2022-02-17).

[imz@ later extended this patch with a comment in backends/apt/meson.build
at this place so that the patch context verifies that we have c++17 there.]

And this value is valid for the compilers in sisyphus, p11, p10, p9,
{sisyphus,p10,p9}_e2k.

If one needs to set it again by force, it'd be nicer to pass it as
-Dcpp_std=c++17 to meson or as part of apt_dep in meson.build
(emulating a pkg-config for APT, which itself would be an even better
solution).

(BTW, the current e2k -Wno-error flag doesn't deprive us of the
interesting fatal compiler errors there as long as individual
-Werror=* are used (in meson.build) instead of -Werror.  Otherwise,
one could put %%remove_optflags -Wno-error in the spec-file.)

(cherry picked from commit 9890681beac9615ec30a941e304acecb15af4c77)
(= 1.2.5.0.0.35-alt1~4)

* * *

Notes on finding a commit like this to pick from the history:

git log --oneline --first-parent gears/sisyphus -i -E --grep='e2k|elbrus'

("tig" can also be used in place of "git log --oneline".)

--first-parent is to look just into the most "up-to-date" (enhanced)
revision of the ALT's patch series, which I regularly rebase onto a new
upstream version (or rewrite to refactor) and then merge as the first parent
into the published branch (sisyphus) before preparing a new release
(according to my maintenance scheme). (Merging as the first parent is
for more meaningful info in "git blame [--first-parent]".)

As for when the change in question initially appeared (in the history
of package releases): this can be learned from the "(cherry picked ...)"
line in the commit message in the rebased patch series, which gives
the ID (hash) of the initial commit.

Note that if searching more branches at once (e.g., with --remotes=...,
e.g., all my work-in-progress branches), the results will be a bigger
mess due to possible repetition of the same patch; that's because
in each branch, the last merged revision of the rebased patch series
may be different, and so several commits bearing the patch in question
will show up in the results.

Another kind of search for a patch through the history--to find the
earliest (initial) appearance of a patch--must necessarily be done with
a command that doesn't restrict the search to --first-parent; for not
having a better idea:

git log --oneline --grep=override --reverse --remotes=@ALT ^upstream/main

("tig" can also be used in place of "git log --oneline".)