for e2k: fix passing a g_autofree_EDG(gchar) to a printf-like function
In other cases, it must be converted automatically, but here the
signature of the function does not specify the type of the arg,
so it is not coerced.

On non-e2k, this is just a char-pointer with a cleanup-fucntion,
so it compiles fine. On e2k, we have a class that does the autocleanup
(the cleanup attribute is not supported for C++ yet on e2k).

This works around the following problem with autofree_EDG(gchar):

lcc: "../backends/apt/apt-job.cpp", line 424: warning #1290: a class type that
          is not trivially copyable passed through ellipsis
                 package_id, local_error->message);
                 ^
 in expansion of macro "g_warning" at line 423
          g_warning ("package_id %s invalid and cannot be processed: %s",
          ^

lcc: "../backends/apt/apt-job.cpp", line 424: error #181: argument is
          incompatible with corresponding format string conversion (expected
          type "char *" but argument has type
          "EDG_CLEANUP_HELPER<gchar, void, &g_autoptr_cleanup_generic_gfree>")
          [-Werror=format]
                 package_id, local_error->message);
                 ^
 in expansion of macro "g_warning" at line 423
          g_warning ("package_id %s invalid and cannot be processed: %s",
          ^

Follows-up (where the code in question appeared): v1.2.6~6
d9697a6cd (apt: Implement support for Packages signal to emit many packages at once, 2022-11-15)

(cherry picked from commit 20128651065fe2a6d0e9f26bf07f0641177a7915)
(= 1.2.6-alt1~1)