runTransaction: repeat DistUpgrade solution exactly without extra installations
For this, we need to know whether the request to install a package comes from a
ready solution for a DistUpgrade (or any similar action). Then we
don't have to do a recursive MarkInstall() that might find a different
solution (a different set of pkgs to install and remove).

For this, we pass the PkgSolved info in PkgInfo and package IDs
(encoded by =).