runTransaction: avoid extra unwanted removals/installs caused by pkgs that shouldn't cause them
The pkgs that come from a ready solution (say, to DistUpgrade) might
still cause some extra unwanted removals if MarkInstall'd recursively
(on the second pass) before the planned removals are done.

It'd be more robust to do the removals after the first pass, but PK
structures the actions so that installs and removals are invoked
separately... :( Additional action needs to be introduced.